pyproject.toml
[project]
name = "paperinator"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Raphael Lüthy", email = "raphael.luethy@fhnw.ch" }
]
dependencies = [
    "google-generativeai>=0.8.4",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/paperinator"]

[tool.ruff]
line-length = 80

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true

[tool.pyright]
include = ["src/*"]
exclude = ["src/test/*"]
venvPath = ".venv"
venv = ".venv"
